requirement.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
celery
django-celery-beat
redis
django-storages[boto3]
boto3
django-filter
django-environ